+++ clients/registry-amqp/src/main/java/org/eclipse/hono/client/registry/amqp/ProtonBasedDeviceRegistrationClient.java
@@ -41,6 +41,7 @@ import org.eclipse.hono.util.MessageHelper;
 import org.eclipse.hono.util.RegistrationAssertion;
 import org.eclipse.hono.util.RegistrationConstants;
 import org.eclipse.hono.util.RegistrationResult;
+import org.eclipse.hono.util.RequestResponseApiConstants;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -173,7 +174,7 @@ public class ProtonBasedDeviceRegistrationClient extends AbstractRequestResponse
                                     RegistrationConstants.ACTION_ASSERT,
                                     properties,
                                     null,
-                                    RegistrationConstants.CONTENT_TYPE_APPLICATION_JSON,
+                                    RequestResponseApiConstants.CONTENT_TYPE_APPLICATION_JSON,
                                     this::getRequestResponseResult,
                                     span);
                         })