+++ core/src/main/java/org/eclipse/hono/util/CommandEndpoint.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2021 Contributors to the Eclipse Foundation
+ * Copyright (c) 2021, 2022 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -130,6 +130,6 @@ public class CommandEndpoint {
     @JsonIgnore
     public String getFormattedUri(final String deviceId) {
         Objects.requireNonNull(deviceId);
-        return uri.replaceAll("\\{\\{deviceId}}", deviceId);
+        return uri.replace("{{deviceId}}", deviceId);
     }
 }