+++ services/base-jdbc/src/main/java/org/eclipse/hono/service/base/jdbc/store/SQL.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 Contributors to the Eclipse Foundation
+ * Copyright (c) 2020, 2022 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -79,9 +79,9 @@ public final class SQL {
             case "23000": //$FALL-THROUGH$
             case "23505":
                 return Future.failedFuture(new DuplicateKeyException(e));
+            default:
+                return Future.failedFuture(e);
         }
-
-        return Future.failedFuture(e);
     }
 
     /**